+++ py/packages/genkit/tests/genkit/ai/agent_session_stores_test.py
@@ -31,7 +31,7 @@ from genkit._core._typing import (
     SnapshotStatus,
     TextPart,
 )
-from genkit.agent import (
+from genkit.exp.agent import (
     FileSessionStore,
     InMemorySessionStore,
     SessionStore,
