+++ py/packages/genkit/tests/genkit/ai/agent_state_schema_server_test.py
@@ -29,7 +29,7 @@ from genkit._core._typing import (
     SessionState,
     SnapshotStatus,
 )
-from genkit.agent import InMemorySessionStore
+from genkit.exp.agent import InMemorySessionStore
 
 INVALID_ARGUMENT = 'INVALID_ARGUMENT'
