+++ py/samples/agents/basic/03_interrupt_resume_with_store.py
@@ -33,11 +33,11 @@ from genkit_middleware import Middleware, ToolApproval
 from pydantic import BaseModel, Field
 
 from genkit import ToolRequestPart
-from genkit.agent import (
+from genkit.exp import Genkit
+from genkit.exp.agent import (
     AgentFinishReason,
     InMemorySessionStore,
 )
-from genkit.exp import Genkit
 
 
 class TransferInput(BaseModel):
