+++ py/samples/agents/basic/06_define_custom_agent.py
@@ -28,7 +28,8 @@ from __future__ import annotations
 from genkit_google_genai import GoogleAI
 
 from genkit import ActionRunContext, FinishReason, Message
-from genkit.agent import (
+from genkit.exp import Genkit
+from genkit.exp.agent import (
     AgentFinishReason,
     AgentInput,
     AgentResult,
@@ -38,7 +39,6 @@ from genkit.agent import (
     TurnContext,
     TurnResult,
 )
-from genkit.exp import Genkit
 
 ai = Genkit(plugins=[GoogleAI()])
 store = InMemorySessionStore()
