+++ py/samples/agents/basic/07_artifacts_custom_patch.py
@@ -33,7 +33,8 @@ from genkit_google_genai import GoogleAI
 from pydantic import BaseModel, Field
 
 from genkit import ActionRunContext, FinishReason, Message, Part, TextPart
-from genkit.agent import (
+from genkit.exp import Genkit
+from genkit.exp.agent import (
     AgentFinishReason,
     AgentInput,
     AgentResult,
@@ -44,7 +45,6 @@ from genkit.agent import (
     TurnContext,
     TurnResult,
 )
-from genkit.exp import Genkit
 
 ai = Genkit(plugins=[GoogleAI()])
 store = InMemorySessionStore()
