+++ py/samples/agents/basic/10_abort.py
@@ -29,8 +29,8 @@ import asyncio
 from genkit_google_genai import GoogleAI
 
 from genkit import GenkitError, ToolRunContext
-from genkit.agent import InMemorySessionStore
 from genkit.exp import Genkit
+from genkit.exp.agent import InMemorySessionStore
 
 ai = Genkit(plugins=[GoogleAI()])
 store = InMemorySessionStore()
