+++ py/samples/agents/basic/11_write_artifact_tool.py
@@ -27,8 +27,8 @@ from __future__ import annotations
 from genkit_google_genai import GoogleAI
 from genkit_middleware import Artifacts, Middleware
 
-from genkit.agent import InMemorySessionStore
 from genkit.exp import Genkit
+from genkit.exp.agent import InMemorySessionStore
 
 ai = Genkit(plugins=[GoogleAI(), Middleware()])
 store = InMemorySessionStore()
