+++ py/samples/agents/basic/12_abort_client_and_server.py
@@ -32,8 +32,8 @@ import asyncio
 from genkit_google_genai import GoogleAI
 
 from genkit import GenkitError, ToolRunContext
-from genkit.agent import InMemorySessionStore, SnapshotStatus
 from genkit.exp import Genkit
+from genkit.exp.agent import InMemorySessionStore, SnapshotStatus
 
 ai = Genkit(plugins=[GoogleAI()])
 store = InMemorySessionStore()
