+++ py/samples/agents/basic/15_branching.py
@@ -34,8 +34,8 @@ from __future__ import annotations
 from genkit_google_genai import GoogleAI
 
 from genkit import GenkitError
-from genkit.agent import InMemorySessionStore
 from genkit.exp import Genkit
+from genkit.exp.agent import InMemorySessionStore
 
 ai = Genkit(plugins=[GoogleAI()])
 # reject_ambiguous_session makes a session-id lookup over a forked history raise
