+++ py/samples/agents/basic/16_time_travel_artifacts.py
@@ -32,8 +32,8 @@ from __future__ import annotations
 from genkit_google_genai import GoogleAI
 from genkit_middleware import Artifacts, Middleware
 
-from genkit.agent import InMemorySessionStore
 from genkit.exp import Genkit
+from genkit.exp.agent import InMemorySessionStore
 
 ai = Genkit(plugins=[GoogleAI(), Middleware()])
