+++ py/samples/agents/basic/17_client_transform.py
@@ -36,7 +36,8 @@ from __future__ import annotations
 from genkit_google_genai import GoogleAI
 
 from genkit import ActionRunContext, FinishReason, Message, Part, TextPart
-from genkit.agent import (
+from genkit.exp import Genkit
+from genkit.exp.agent import (
     AgentFinishReason,
     AgentInput,
     AgentResult,
@@ -47,7 +48,6 @@ from genkit.agent import (
     TurnContext,
     TurnResult,
 )
-from genkit.exp import Genkit
 
 ai = Genkit(plugins=[GoogleAI()])
