+++ py/samples/agents/testapp/trip_planner_agent.py
@@ -30,7 +30,7 @@ from _ai import ai
 from pydantic import BaseModel
 
 from genkit import ActionRunContext
-from genkit.agent import FileSessionStore
+from genkit.exp.agent import FileSessionStore
 
 
 class CityInput(BaseModel):
